Tests/LibWeb/Text/expected/SVG/svg-getScreenCTM.txt

a: 2, b: 0, c: 0, d: 2, e: 108, f: 158
(inversed) a: 0.5, b: 0, c: 0, d: 0.5, e: -54, f: -79
